# internal/panel/api/openapi.yaml
openapi: 3.0.3
info:
  title: Kraken Panel API
  description: >
    Control-plane REST API for the Kraken self-hosted game-server platform:
    authentication, the game-spec catalog, server lifecycle, files, backups,
    cron schedules, node registry, Agent mTLS enrollment, users/roles, and the
    audit log. All endpoints are under /api/v1 unless noted. Authenticated
    endpoints use a bearer session token from POST /auth/login.
  version: "1.0.0"
servers:
  - url: /api/v1
    description: Panel API (relative to the Panel host)
...
security:
  - bearerAuth: []
paths:
...
  /servers/{id}/backups:
    parameters: [ { $ref: '#/components/parameters/ServerId' } ]
    get:
      tags: [Backups]
      summary: List backups
      description: >
        Also answers for a retired server — keeping its archives listable is
        the point of retiring — from the node it was retired from. When that
        node no longer exists the answer is 404 `not_found`: its archives went
        with it.
      responses:
        '200':
          description: Backups
          content:
            application/json:
              schema:
                type: object
                properties:
                  backups:
                    type: array
                    items: { $ref: '#/components/schemas/Backup' }
        '404': { $ref: '#/components/responses/AgentNotFound' }
        '409': { $ref: '#/components/responses/AgentConflict' }
        '500': { $ref: '#/components/responses/AgentError' }
        '503': { $ref: '#/components/responses/NodeUnreachable' }
    post:
      tags: [Backups]
      summary: Create a backup
      requestBody:
        content:
          application/json:
            schema: { type: object, properties: { name: { type: string } } }
      responses:
        '202':
          description: Archiving started; poll the list for the backup to reach ready
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Backup' }
        '404': { $ref: '#/components/responses/AgentNotFound' }
        '409': { $ref: '#/components/responses/AgentConflict' }
        '500': { $ref: '#/components/responses/AgentError' }
        '503': { $ref: '#/components/responses/NodeUnreachable' }
...
  /servers/{id}/backups/{backupId}:
    parameters:
      - { $ref: '#/components/parameters/ServerId' }
      - { name: backupId, in: path, required: true, schema: { type: string } }
    delete:
      tags: [Backups]
      summary: Delete a backup
      responses:
        '204': { description: Deleted }
        '404': { $ref: '#/components/responses/AgentNotFound' }
        '409': { $ref: '#/components/responses/AgentConflict' }
        '500': { $ref: '#/components/responses/AgentError' }
        '503': { $ref: '#/components/responses/NodeUnreachable' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Session token from POST /auth/login
  parameters:
    ServerId:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
...
    AgentNotFound:
      description: >
        No such server, or one this user may not reach (existence is not
        revealed) — or the path or backup does not exist on the node
        (`code: not_found`).
      content:
        application/json: { schema: { $ref: '#/components/schemas/Error' } }
    AgentConflict:
      description: >
        The node refused the operation, and `code` says why: `file_in_use`
        (another process holds the file — usually the server's own game, still
        running), `node_refused` (the node's filesystem denied it) or
        `already_exists`. `error` carries the node's own words. On an
        operation that writes the server's files, `server_restoring` means a
        backup restore holds them and `server_busy` that a retire, revive or
        delete holds the server; nothing was sent to the node. On a retired
        server every file and backup route but the backup list answers
        `server_retired` (revive it first).
      content:
        application/json: { schema: { $ref: '#/components/schemas/Error' } }
...
    AgentError:
      description: >
        The node reported a failure it did not classify (`code: node_error`);
        `error` carries its message.
      content:
        application/json: { schema: { $ref: '#/components/schemas/Error' } }
    NodeUnreachable:
      description: >
        The panel could not reach the node's agent, or it did not answer in
        time (`code: node_unreachable`).
      content:
        application/json: { schema: { $ref: '#/components/schemas/Error' } }
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          description: The human-readable reason, as the UI shows it.
        code:
          type: string
          description: >
            A stable, machine-readable reason, present on the failures a
            client may want to branch on — branch on this, never on `error`.
            A failed Agent operation carries one of node_unreachable,
            not_found, bad_path, file_in_use, node_refused, already_exists,
            install_running or
            node_error; a settings save whose config could not be rendered
            carries config_apply_failed. A backup restore's refusals carry
            server_restoring (a restore holds the server's files, so the
            write, start or delete was refused), restore_in_progress (a second
            restore) or server_busy (a start, restart or reinstall still holds
            the server). The retire model (#360) adds server_retired (the
            server is retired: nothing boots it, writes its files or backs it
            up — revive it first), server_not_retired (revive and permanent
            delete are only for a retired server), removal_pending (a removal
            for the server is still owed to a node) and server_busy for a retire
            or revive in progress.
...
    Backup:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        size: { type: integer }
        created_ms: { type: integer, format: int64 }
        state: { type: string, enum: [pending, ready, failed], description: archive lifecycle — backups run asynchronously }
        replication: { type: string, enum: ['', pending, done, failed], description: off-node (SFTP) mirror status; empty when replication isn't configured }
        error: { type: string, description: "why the backup failed (state=failed), or a degraded-capture note on a ready archive; omitted when clean" }
...